Move exit manager to its own loop
